src/modules/reminder_module.py
from datetime import datetime
from typing import Dict, Any
import asyncio
from dateutil import parser
from .base_module import BaseModule

class ReminderModule(BaseModule):
    """Module for setting and managing reminders."""
    
    def __init__(self):
        """Initialize the reminder module."""
        super().__init__()
        self.reminders = []
    
    def get_description(self) -> str:
        """Get module description."""
        return "Sets and manages reminders for various tasks and events"
    
    def get_required_parameters(self) -> Dict[str, str]:
        """Get required parameters."""
        return {
            "task": "The task or event to be reminded about",
            "time": "When the reminder should trigger (datetime or natural language)",
        }
    
    async def execute(self, parameters: Dict[str, Any]) -> Dict[str, Any]:
        """Execute the reminder task."""
        if not self.validate_parameters(parameters):
            return self.format_error("Missing required parameters")
        
        try:
            # Parse the time parameter
            reminder_time = self._parse_time(parameters["time"])
            if reminder_time < datetime.now():
                return self.format_error("Reminder time must be in the future")
            
            # Create the reminder
            reminder = {
                "task": parameters["task"],
                "time": reminder_time,
                "created_at": datetime.now()
            }
            
            # Add to reminders list
            self.reminders.append(reminder)
            
            # Schedule the reminder notification
            asyncio.create_task(self._schedule_reminder(reminder))
            
            return self.format_success({
                "message": f"Reminder set for {reminder_time.strftime('%Y-%m-%d %H:%M:%S')}",
                "task": parameters["task"]
            })
            
        except Exception as e:
            return self.format_error(f"Failed to set reminder: {str(e)}")
    
    def _parse_time(self, time_str: str) -> datetime:
        """
        Parse a time string into a datetime object.
        
        Args:
            time_str (str): Time string in natural language or datetime format
            
        Returns:
            datetime: Parsed datetime object
        """
        try:
            return parser.parse(time_str)
        except Exception as e:
            raise ValueError(f"Could not parse time: {time_str}")
    
    async def _schedule_reminder(self, reminder: Dict[str, Any]):
        """
        Schedule a reminder to be triggered at the specified time.
        
        Args:
            reminder (Dict[str, Any]): The reminder to schedule
        """
        now = datetime.now()
        delay = (reminder["time"] - now).total_seconds()
        
        if delay > 0:
            await asyncio.sleep(delay)
            # TODO: Implement actual notification system
            print(f"\nREMINDER: {reminder['task']}") 