requirements.txt
python-dotenv==1.0.1
openai==1.66.2
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pytest==8.0.0
python-dateutil==2.8.2
rich==13.7.1
tiktoken==0.5.2
aiohttp==3.9.3
anyio==4.8.0 